80. Strcuts completed
Completed course, understood building a structure, how it was utilized in the game wouldn't of been my first guess.

Modified the HiddenWordList.h to onlyl have 5 options. This seemed to resolve the long compile times that occurred even when just editing a comment.  Will look at this more when back in office.

Tested code
- Compiled successfully
- Launched Play in Editor successfully
- Feedback for if player enters in to small of word
- Feedback for if player enters in non isogram word
- Feedback for if player makes wrong guess / loses guess
- Player is shown how many Bulls and Cows from guess
- Player lose condition
- Player win condition
- Play again